Use contacts API v2
This commit uses the new contacts API to get all the contacts.
Retrieving a single contact is no longer possible.
Accepting and rejecting contact requests will need to be updated
as well (it still uses the old API).

New types were written for this update. The new types are in the
`types` directory must be imported using deep imports
(`import {Contact} from "skype-http/types/contact";`).
The old types are still in `lib/interfaces` for the moment.