Add support for login with email
This commit adds a step during the login operation to fetch the
username from the Skype DB instead of just passing the login. This
means that it is now possible to use email or usernames with mixed case
to login.

It also adds try/catch blocks around the read operation when
deserializing the results, registers the `GroupConversation` MRI type,
renames `types/profile` to `types/contact-profile` and fixes some
errors with the interfaces.

- Closes ocilo/skype-http#58